partial feedback: wasm_dummy removal, move from 24-host module, gas meter fix
